chromium-wayland_%.bbappend: rename it and update OpenCDM to match v65
The chromium for wayland recipe name has changed when chromium was updated
from v53 to v65.

Currently only External Clear Key is supported.
Playready support, and the possibility to do a debug build of OpenCDM
while having a release build of Chromium are TBD (to be forward ported
from morty).

Signed-off-by: Andrey Konovalov <andrey.konovalov@linaro.org>